+++ pymatgen/io/abinitio/abiobjects.py
@@ -14,8 +14,9 @@ from pprint import pformat
 from monty.design_patterns import singleton
 from monty.collections import AttrDict
 from pymatgen.core.design_patterns import Enum
-from pymatgen.serializers.json_coders import PMGSONable
+from pymatgen.serializers.json_coders import PMGSONable, pmg_serialize
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
+from monty.json import MontyEncoder, MontyDecoder
 
 
 def contract(s):
@@ -76,7 +77,7 @@ MANDATORY = MandatoryVariable()
 DEFAULT = DefaultVariable()
 
 
-class SpinMode(collections.namedtuple('SpinMode', "mode nsppol nspinor nspden"), AbivarAble):
+class SpinMode(collections.namedtuple('SpinMode', "mode nsppol nspinor nspden"), AbivarAble, PMGSONable):
     """
     Different configurations of the electron density as implemented in abinit:
     One can use as_spinmode to construct the object via SpinMode.as_spinmode
@@ -106,6 +107,15 @@ class SpinMode(collections.namedtuple('SpinMode', "mode nsppol nspinor nspden"),
             "nspinor": self.nspinor,
             "nspden": self.nspden,
         }
+
+    @pmg_serialize
+    def as_dict(self):
+        return {k: getattr(self, k) for k in self._fields}
+
+    @classmethod
+    def from_dict(cls, d):
+        return cls(**{k: d[k] for k in d if k in cls._fields})
+
 
 # An handy Multiton
 _mode2spinvars = {
@@ -215,7 +225,7 @@ class Smearing(AbivarAble, PMGSONable):
         return Smearing(d["occopt"], d["tsmear"])
 
 
-class ElectronsAlgorithm(dict, AbivarAble):
+class ElectronsAlgorithm(dict, AbivarAble, PMGSONable):
     """Variables controlling the SCF/NSCF algorithm."""
     # None indicates that we use abinit defaults.
     _DEFAULT = dict(
@@ -233,8 +243,21 @@ class ElectronsAlgorithm(dict, AbivarAble):
     def to_abivars(self):
         return self.copy()
 
+    def as_dict(self):
+        d = self.copy()
+        d["@module"] = self.__class__.__module__
+        d["@class"] = self.__class__.__name__
+        return d
 
-class Electrons(AbivarAble):
+    @classmethod
+    def from_dict(cls, d):
+        d = d.copy()
+        d.pop("@module", None)
+        d.pop("@class", None)
+        return cls(**d)
+
+
+class Electrons(AbivarAble, PMGSONable):
     """The electronic degrees of freedom"""
     def __init__(self, spin_mode="polarized", smearing="fermi_dirac:0.1 eV",
                  algorithm=None, nband=None, fband=None, charge=0.0, comment=None):  # occupancies=None,
@@ -268,18 +291,31 @@ class Electrons(AbivarAble):
     def nspden(self):
         return self.spin_mode.nspden
 
-    #@property
-    #def as_dict(self):
-    #    "json friendly dict representation"
-    #    d = {}
-    #    d["@module"] = self.__class__.__module__
-    #    d["@class"] = self.__class__.__name__
-    #    raise NotImplementedError("")
-    #    return d
+    def as_dict(self):
+        "json friendly dict representation"
+        d = {}
+        d["@module"] = self.__class__.__module__
+        d["@class"] = self.__class__.__name__
+        d["spin_mode"] = self.spin_mode.as_dict()
+        d["smearing"] = self.smearing.as_dict()
+        d["algorithm"] = self.algorithm.as_dict() if self.algorithm else None
+        d["nband"] = self.nband
+        d["fband"] = self.fband
+        d["charge"] = self.charge
+        d["comment"] = self.comment
+        return d
 
-    #@staticmethod
-    #def from_dict(d):
-    #    raise NotImplementedError("")
+    @classmethod
+    def from_dict(cls, d):
+        d = d.copy()
+        d.pop("@module", None)
+        d.pop("@class", None)
+        dec = MontyDecoder()
+        d["spin_mode"] = dec.process_decoded(d["spin_mode"])
+        d["smearing"] = dec.process_decoded(d["smearing"])
+        d["algorithm"] = dec.process_decoded(d["algorithm"]) if d["algorithm"] else None
+        return cls(**d)
+
 
     def to_abivars(self):
         abivars = self.spin_mode.to_abivars()
@@ -296,11 +332,11 @@ class Electrons(AbivarAble):
         if self.algorithm:
             abivars.update(self.algorithm)
 
-        abivars["#comment"] = self.comment
+        #abivars["#comment"] = self.comment
         return abivars
 
 
-class KSampling(AbivarAble):
+class KSampling(AbivarAble, PMGSONable):
     """
     Input variables defining the K-point sampling.
     """
@@ -352,6 +388,14 @@ class KSampling(AbivarAble):
 
         self.mode = mode
         self.comment = comment
+
+        self.num_kpts = num_kpts
+        self.kpts = kpts
+        self.kpt_shifts = kpt_shifts
+        self.kpts_weights = kpts_weights
+        self.use_symmetries = use_symmetries
+        self.use_time_reversal = use_time_reversal
+        self.chksymbreak = chksymbreak
 
         abivars = {}
 
@@ -406,7 +450,7 @@ class KSampling(AbivarAble):
             raise ValueError("Unknown mode %s" % mode)
 
         self.abivars = abivars
-        self.abivars["#comment"] = comment
+        #self.abivars["#comment"] = comment
 
     @property
     def is_homogeneous(self):
@@ -463,7 +507,7 @@ class KSampling(AbivarAble):
         Convenient static constructor for an automatic Monkhorst-Pack mesh.
 
         Args:
-            structure: pymatgen structure object.
+            structure: :class:`Structure` object.
             ngkpt: Subdivisions N_1, N_2 and N_3 along reciprocal lattice vectors.
             use_symmetries: Use spatial symmetries to reduce the number of k-points.
             use_time_reversal: Use time-reversal symmetry to reduce the number of k-points.
@@ -493,7 +537,7 @@ class KSampling(AbivarAble):
         Static constructor for path in k-space.
 
         Args:
-            structure: pymatgen structure.
+            structure: :class:`Structure` object.
             kpath_bounds: List with the reduced coordinates of the k-points defining the path.
             ndivsm: Number of division for the smallest segment.
             comment: Comment string.
@@ -523,7 +567,7 @@ class KSampling(AbivarAble):
     @classmethod
     def path_from_structure(cls, ndivsm, structure):
         """See _path for the meaning of the variables"""
-        return cls._path(ndivsm,  structure=structure, comment="K-path generated automatically from pymatgen structure")
+        return cls._path(ndivsm,  structure=structure, comment="K-path generated automatically from structure")
 
     @classmethod
     def explicit_path(cls, ndivsm, kpath_bounds):
@@ -574,7 +618,7 @@ class KSampling(AbivarAble):
         #    num_div = [i + i % 2 for i in num_div]
         #    style = Kpoints.modes.monkhorst
 
-        comment = "pymatgen generated KPOINTS with grid density = " + "{} / atom".format(kppa)
+        comment = "abinitio generated KPOINTS with grid density = " + "{} / atom".format(kppa)
 
         shifts = np.reshape(shifts, (-1, 3))
 
@@ -586,6 +630,23 @@ class KSampling(AbivarAble):
     def to_abivars(self):
         return self.abivars
 
+    def as_dict(self):
+        enc = MontyEncoder()
+        return {'mode': self.mode, 'comment': self.comment, 'num_kpts': self.num_kpts,
+                'kpts': enc.default(np.array(self.kpts)), 'kpt_shifts': self.kpt_shifts,
+                'kpts_weights': self.kpts_weights, 'use_symmetries': self.use_symmetries,
+                'use_time_reversal': self.use_time_reversal, 'chksymbreak': self.chksymbreak,
+                '@module': self.__class__.__module__, '@class': self.__class__.__name__}
+
+    @classmethod
+    def from_dict(cls, d):
+        d = d.copy()
+        d.pop('@module', None)
+        d.pop('@class', None)
+        dec = MontyDecoder()
+        d['kpts'] = dec.process_decoded(d['kpts'])
+        return cls(**d)
+
 
 class Constraints(AbivarAble):
     """This object defines the constraints for structural relaxation"""
@@ -593,7 +654,7 @@ class Constraints(AbivarAble):
         raise NotImplementedError("")
 
 
-class RelaxationMethod(AbivarAble):
+class RelaxationMethod(AbivarAble, PMGSONable):
     """
     This object stores the variables for the (constrained) structural optimization
     ionmov and optcell specify the type of relaxation.
@@ -690,6 +751,20 @@ class RelaxationMethod(AbivarAble):
             })
 
         return out_vars
+
+    def as_dict(self):
+        d = dict(self._default_vars)
+        d['@module'] = self.__class__.__module__
+        d['@class'] = self.__class__.__name__
+        return d
+
+    @classmethod
+    def from_dict(cls, d):
+        d = d.copy()
+        d.pop('@module', None)
+        d.pop('@class', None)
+
+        return cls(**d)
 
 
 class PPModel(AbivarAble, PMGSONable):