+++ pymatgen/io/vaspio_set.py
@@ -31,7 +31,6 @@ import numpy as np
 
 from monty.serialization import loadfn
 
-from pymatgen.io.cifio import CifWriter
 from pymatgen.io.vaspio.vasp_input import Incar, Poscar, Potcar, Kpoints
 from pymatgen.io.vaspio.vasp_output import Vasprun, Outcar
 from pymatgen.serializers.json_coders import PMGSONable
@@ -125,16 +124,23 @@ class AbstractVaspInputSet(six.with_metaclass(abc.ABCMeta, PMGSONable)):
         Returns:
             dict of {filename: file_as_string}, e.g., {'INCAR':'EDIFF=1e-4...'}
         """
-        d = {'INCAR': self.get_incar(structure),
-             'KPOINTS': self.get_kpoints(structure),
+        kpoints = self.get_kpoints(structure)
+        incar = self.get_incar(structure)
+        if np.product(kpoints.kpts) < 4 and incar.get("ISMEAR", 0) == -5:
+            incar["ISMEAR"] = 0
+
+        d = {'INCAR': incar,
+             'KPOINTS': kpoints,
              'POSCAR': self.get_poscar(structure)}
+
         if generate_potcar:
             d['POTCAR'] = self.get_potcar(structure)
         else:
             d['POTCAR.spec'] = "\n".join(self.get_potcar_symbols(structure))
         return d
 
-    def write_input(self, structure, output_dir, make_dir_if_not_present=True):
+    def write_input(self, structure, output_dir,
+                    make_dir_if_not_present=True, include_cif=False):
         """
         Writes a set of VASP input to a directory.
 
@@ -145,11 +151,17 @@ class AbstractVaspInputSet(six.with_metaclass(abc.ABCMeta, PMGSONable)):
             make_dir_if_not_present (bool): Set to True if you want the
                 directory (and the whole path) to be created if it is not
                 present.
+            include_cif (bool): Whether to write a CIF file in the output
+                directory for easier opening by VESTA.
         """
         if make_dir_if_not_present and not os.path.exists(output_dir):
             os.makedirs(output_dir)
         for k, v in self.get_all_vasp_input(structure).items():
             v.write_file(os.path.join(output_dir, k))
+            if k == "POSCAR" and include_cif:
+                v.structure.to(
+                    filename=os.path.join(output_dir,
+                                          "%s.cif" % v.structure.formula))
 
 
 class DictVaspInputSet(AbstractVaspInputSet):
@@ -203,7 +215,6 @@ class DictVaspInputSet(AbstractVaspInputSet):
         reduce_structure (None/str): Before generating the input files,
             generate the reduced structure. Default (None), does not
             alter the structure. Valid values: None, "niggli", "LLL"
-
     """
 
     def __init__(self, name, config_dict, hubbard_off=False,
@@ -490,7 +501,15 @@ class MITNEBVaspInputSet(DictVaspInputSet):
             structures.append(s)
         return structures
 
-    def write_input(self, structures, output_dir, make_dir_if_not_present=True,
+    def write_input(self, structure, output_dir,
+                    make_dir_if_not_present=True, include_cif=False):
+        """
+        The default write_input is for writing initial structures / end point
+        relaxations. This ensures that
+        """
+
+    def write_input(self, structures, output_dir,
+                    make_dir_if_not_present=True,
                     write_cif=False):
         """
         NEB inputs has a special directory structure where inputs are in 00,
@@ -1062,7 +1081,8 @@ class MPNonSCFVaspInputSet(MPStaticVaspInputSet):
         """
         if self.mode == "Line":
             kpath = HighSymmKpath(structure)
-            cart_k_points, k_points_labels = kpath.get_kpoints(line_density=self.kpoints_line_density)
+            cart_k_points, k_points_labels = kpath.get_kpoints(
+                line_density=self.kpoints_line_density)
             frac_k_points = [kpath._prim_rec.get_fractional_coords(k)
                              for k in cart_k_points]
             return Kpoints(comment="Non SCF run along symmetry lines",
@@ -1391,9 +1411,6 @@ def batch_write_vasp_input(structures, vasp_input_set, output_dir,
         if sanitize:
             s = s.copy(sanitize=True)
         vasp_input_set.write_input(
-            s, dirname, make_dir_if_not_present=make_dir_if_not_present
+            s, dirname, make_dir_if_not_present=make_dir_if_not_present,
+            include_cif=include_cif
         )
-        if include_cif:
-            writer = CifWriter(s)
-            writer.write_file(os.path.join(
-                dirname, "{}_{}.cif".format(formula, i)))